requirements.txt
playwright==1.40.0
beautifulsoup4==4.12.2
pandas==2.1.4
streamlit==1.29.0
requests==2.31.0
lxml==4.9.3
plotly==5.18.0
networkx==3.2.1
feedparser==6.0.11
supabase==2.3.0
python-dotenv==1.0.0
